requirements.txt
openai>=1.30.0
azure-identity>=1.16.0
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.7.0
pydantic>=2.7.0
dnspython>=2.6.0
python-nmap>=0.7.1
pyOpenSSL>=24.0.0
cryptography>=42.0.0
click>=8.1.0
urllib3>=2.0.0
